Implement dynamic validation system and improve error handling
- Add min_length and max_length fields to AttributeType entity for configurable validation
- Create IdentityValidationService for dynamic validation based on attribute type configuration
- Remove hardcoded @Size validation from CreateIdentityRequest DTO
- Add ValidationException and proper error handling in GlobalExceptionHandler
- Create database migrations V4 and V5 for validation fields and core attributes
- Move validation rules from code to database configuration
- Return proper HTTP 400 responses with structured validation errors instead of 500 errors
- Support configurable validation per identity type and attribute

+++ src/main/resources/db/migration/V2__create_identity_type_system.sql
@@ -34,7 +34,7 @@ CREATE TABLE identity_attribute_values (
     attribute_type_id VARCHAR(36) NOT NULL,
     string_value TEXT,
     integer_value BIGINT,
-    decimal_value DECIMAL(19,4),
+    decimal_value DOUBLE PRECISION,
     boolean_value BOOLEAN,
     date_value TIMESTAMP WITH TIME ZONE,
     created_at TIMESTAMP WITH TIME ZONE NOT NULL,
